+++ finance/src/sgpp/finance/application/HullWhiteSolver.cpp
@@ -56,7 +56,7 @@ HullWhiteSolver::~HullWhiteSolver() {
   }
 }
 
-void HullWhiteSolver::constructGrid(base::BoundingBox& BoundingBox, int level) {
+void HullWhiteSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
   this->levels = level;
