+++ pde/src/sgpp/pde/application/HeatEquationSolverWithStretching.cpp
@@ -31,7 +31,7 @@ HeatEquationSolverWithStretching::~HeatEquationSolverWithStretching() {
   }
 }
 
-void HeatEquationSolverWithStretching::constructGrid(base::Stretching& stretching, int level) {
+void HeatEquationSolverWithStretching::constructGrid(base::Stretching& stretching, size_t level) {
   this->dim = stretching.getDimensions();
   this->levels = level;
 
@@ -45,7 +45,7 @@ void HeatEquationSolverWithStretching::constructGrid(base::Stretching& stretchin
   this->bGridConstructed = true;
 }
 
-void HeatEquationSolverWithStretching::constructGrid(base::BoundingBox& BoundingBox, int level) {
+void HeatEquationSolverWithStretching::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   std::cout << "I'm not supposed to be here, me is constructGrid\n";
 }
 
@@ -231,14 +231,14 @@ void HeatEquationSolverWithStretching::initScreen() {
 }
 
 void HeatEquationSolverWithStretching::printGrid(base::DataVector& alpha,
-                                                 double PointesPerDimension,
+                                                 size_t PointesPerDimension,
                                                  std::string tfilename) const {
   base::GridPrinterForStretching myPrinter(*this->myGrid);
-  myPrinter.printGrid(alpha, tfilename, static_cast<size_t>(PointesPerDimension));
+  myPrinter.printGrid(alpha, tfilename, PointesPerDimension);
 }
 
 void HeatEquationSolverWithStretching::printGridDomain(base::DataVector& alpha,
-                                                       double PointesPerDimension,
+                                                       size_t PointesPerDimension,
                                                        base::BoundingBox& GridArea,
                                                        std::string tfilename) const {
   throw base::application_exception(
@@ -247,12 +247,11 @@ void HeatEquationSolverWithStretching::printGridDomain(base::DataVector& alpha,
 }
 
 void HeatEquationSolverWithStretching::printGridDomainStretching(base::DataVector& alpha,
-                                                                 double PointesPerDimension,
+                                                                 size_t PointesPerDimension,
                                                                  base::Stretching& GridArea,
                                                                  std::string tfilename) const {
   base::GridPrinterForStretching myPrinter(*this->myGrid);
-  myPrinter.printGridDomainStretching(alpha, tfilename, GridArea,
-                                      static_cast<size_t>(PointesPerDimension));
+  myPrinter.printGridDomainStretching(alpha, tfilename, GridArea, PointesPerDimension);
 }
 
 void HeatEquationSolverWithStretching::printSparseGrid(base::DataVector& alpha,
